# api/openapi/tasks.yaml
openapi: 3.0.3
info:
  title: Tasks API
  version: 1.0.0

# Task 任务管理 API
paths:
...
  /api/v1/tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTask
      summary: 获取任务详情
      parameters:
        - $ref: 'common.yaml#/components/parameters/IdParam'
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
    put:
      tags: [Tasks]
      operationId: updateTask
      summary: 更新任务
      parameters:
        - $ref: 'common.yaml#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: 'common.yaml#/components/responses/BadRequest'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: 删除任务
      parameters:
        - $ref: 'common.yaml#/components/parameters/IdParam'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: 'common.yaml#/components/schemas/MessageResponse'
        '404':
          $ref: 'common.yaml#/components/responses/NotFound'
...
components:
  schemas:
    Task:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
        status:
          type: string
          description: 任务状态（pending=待处理, in_progress=处理中, completed=已完成, failed=已失败, cancelled=已取消）
          enum: [pending, in_progress, completed, failed, cancelled]
        prompt:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        parent_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed, cancelled]
        prompt:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
...